legacy_v1/scripts/data-pipeline/data-pipeline-automation/requirements.txt
google-cloud-bigquerygoogle-cloud-storageos